# backend/docs/swagger.yaml
basePath: /
definitions:
...
  controllers.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  controllers.UserSerializer:
    properties:
      email:
        type: string
      first_name:
        type: string
      id:
        type: integer
      last_name:
        type: string
    type: object
...
  models.User:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      email:
        type: string
      first_name:
        type: string
      first_session:
        type: boolean
      id:
        type: integer
      is_deleted:
        type: boolean
      last_name:
        type: string
      password:
        minLength: 8
        type: string
      updated_at:
        type: string
      verification_code:
        type: string
      verification_code_expiry:
        type: string
      verified:
        type: boolean
    required:
    - email
    - first_name
    - password
    type: object
  models.UserActivity:
    properties:
      date:
        type: string
      id:
        type: integer
      message_count:
        type: integer
      mood:
        type: string
      user_id:
        type: integer
    type: object
...
host: localhost:5050
info:
  contact:
    email: support@swagger.io
    name: Arly API Support
    url: http://www.swagger.io/support
  description: This is the Arly API server documentation
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Arly API
  version: "1.0"
paths:
...
  /api/mood:
    post:
      consumes:
      - application/json
      description: Logs the mood of a user for a given day. If a mood already exists
        for the day, it updates the mood.
      parameters:
      - default: 'Bearer #'
        description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: User activity information
        in: body
        name: activity
        required: true
        schema:
          $ref: '#/definitions/models.UserActivity'
      produces:
      - application/json
      responses:
        "200":
          description: Mood updated successfully
          schema:
            $ref: '#/definitions/models.UserActivity'
        "201":
          description: Mood logged successfully
          schema:
            $ref: '#/definitions/models.UserActivity'
        "400":
          description: Invalid input or missing required fields
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Log a user's mood
      tags:
      - Mood
...
  /api/user:
    get:
      description: Retrieves the details of the currently authenticated user.
      parameters:
      - default: 'Bearer #'
        description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User found
          schema:
            $ref: '#/definitions/controllers.UserSerializer'
        "401":
          description: Unauthorized access
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get the authenticated user's details
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: Updates the details of the currently authenticated user.
      parameters:
      - default: 'Bearer #'
        description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Updated user data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "200":
          description: User updated
          schema:
            $ref: '#/definitions/controllers.UserSerializer'
        "400":
          description: Invalid input or missing required fields
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "401":
          description: Unauthorized access
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update the authenticated user's details
      tags:
      - Users
  /api/user/activities:
    get:
      consumes:
      - application/json
      description: Fetches all activity logs of the authenticated user, optionally
        filtered by a date range.
      parameters:
      - default: 'Bearer #'
        description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Start date in YYYY-MM-DD format
        in: query
        name: start_date
        type: string
      - description: End date in YYYY-MM-DD format
        in: query
        name: end_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of user activities
          schema:
            items:
              $ref: '#/definitions/models.UserActivity'
            type: array
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: No activity found for the user
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: Get the activity history of the authenticated user
      tags:
      - Activity
  /api/user/activity:
    get:
      consumes:
      - application/json
      description: Fetches the activity of the authenticated user for the specified
        date. Defaults to today's activity if no date is provided.
      parameters:
      - default: 'Bearer #'
        description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Date in YYYY-MM-DD format (defaults to today)
        in: query
        name: date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User's activity for the day
          schema:
            $ref: '#/definitions/models.UserActivity'
        "400":
          description: Invalid date format
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "404":
          description: No activity found for the user on the specified date
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/controllers.ErrorResponse'
      summary: Get the activity of the authenticated user for a specific date
      tags:
      - Activity
...
schemes:
- http
swagger: "2.0"